Writer.tmp/backend/requirements.txt
fastapi>=0.111,<1
uvicorn>=0.30,<1
sqlalchemy>=2.0,<3
alembic>=1.13,<2
psycopg[binary]>=3.1,<4
pydantic>=2.0,<3
PyJWT>=2.8,<3
pytest>=8.0,<9
